Upgrader: don't auto-close the zenity dialog when 100% has been reached
We added --auto-close in d6ed609f4785415e4f352c131363899b540f93e7,
as part of #16603, to hide the insensitive "OK" button, but as of 6.8.1
it does not make any difference: that button is displayed regardless
of --auto-close. So --auto-close is not useful anymore and can only
cause trouble.

Indeed, I suspect that --auto-close causes a race condition that occasionally
triggers #20593 when the download has reached 100%:

 - in iteration N-1
   - download_out reaches 100
   - we send 100 to zenity
   - zenity detects this and the code to auto-close is started,
     and starts its race against:

 - in iteration N
   - $zenity_h->pumpable can win the race and still be true,
     if the auto-close code has not finished and zenity has not closed yet
   - by the time we write "100\n" *again* to zenity's stdin,
     it's possible that the auto-close close has finished its job,
     and zenity was closed, so this operation raises a "Broken pipe" exception

refs #20593