Make TCA control the tor-has-bootstrapped flag file (refs: tails#20530)
This drops the (system) tails-wait-until-tor-has-bootstrapped.service
unit, which tails-tor-has-bootstrapped-flag-file.service used to wait
for. Instead TCA monitors Tor's bootstrap status and starts/stops
tails-tor-has-bootstrapped-flag-file.service accordingly.

The reason for this is that we want the flag file (and the connected
tails-tor-has-bootstrapped.target) to be created (and reached) when
are done restarting tor, which TCA does after it sets the time, and
which would break any connections to the tor SocksPort. We have seen
such issues with htpdate when re-run if we reconnect the network:

1. In 10-tor.sh, we restart tor, Tor Connection, and
   htpdate.service. The latter waits for tor to have bootstrapped
   before actually starting htpdate.
2. Since DisableNetwork is not set anymore, tor bootstraps immediately
   and very quickly
3. This creates the flag file that htpdate.service was waiting for, so
   it starts htpdate.
4. Tor Connection sets the clock from the network, then restarts tor,
   which breaks htpdate.