10-tor.sh: don't restart tor (refs: tails#18530)
The restart occurs before we wait for DisableNetwork to be disabled,
so at least for the first network connection the restart seems
useless.

However, on subsequent network connections DisableNetwork can already
be disabled (if an attempt to connect to Tor was made previously) so
then it maybe has an effect. While we could consider keeping
restarting tor in just this case, I'm opting to just drop the restart
completely; we have no indications that there are issues in this area
any more, and, as nickm pointed out [0], the bootstrapping code has
been "seriously rewritten" since we put these hacks in place, so IMHO
we have good reasons to try without.

[0] https://gitlab.torproject.org/tpo/core/tor/-/issues/1247